requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
db.py==0.5.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
numpy==1.25.2
pandas==2.0.3
prettytable==0.7.2
protobuf==4.21.12
pybars3==0.9.7
PyMeta3==0.5.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.6
